+++ plugins/guests/arch/guest.rb
@@ -23,17 +23,16 @@ module VagrantPlugins
 
       def configure_networks(networks)
         networks.each do |network|
-          interfaces.add(network[:interface])
-
-          entry = TemplateRenderer.render("guests/arch/network_#{network[:type]}", :options => network)
+          entry = TemplateRenderer.render("guests/arch/network_#{network[:type]}",
+                                          :options => network)
 
           temp = Tempfile.new("vagrant")
           temp.binmode
           temp.write(entry)
           temp.close
 
-          vm.channel.upload(temp.path, temp.path)
-          vm.channel.sudo("mv #{temp.path} /etc/network.d/interfaces/eth#{network[:interface]}")
+          vm.channel.upload(temp.path, "/tmp/vagrant_network")
+          vm.channel.sudo("mv /tmp/vagrant_network /etc/network.d/interfaces/eth#{network[:interface]}")
           vm.channel.sudo("netcfg interfaces/eth#{network[:interface]}")
         end
       end